調整假日班限制

+++ resources/views/pages/shift-first-edition.blade.php
@@ -643,12 +643,24 @@
                 }
 
                  else if( (array[0]['doc1weekend']<=4) && ( weekday2==6 || weekday2 ==7)  && ( weekday1!=6 && weekday1!=7) ){
-                    dhtmlx.message({ type:"error", text:array[0]['doc1']+"醫生假日班不得少於4" });
-                    console.log("$week1"+weekday1);
-                    console.log("$week1"+weekday2);
+                    var check = confirm(array[0]['doc1']+"醫生假日班不得少於4\n確定要換班嗎?");
+
+                    if(check==true){
+                       save_form();
+                    }
+                    else{
+                        alert("已取消");
+                    }
                 }
                 else if((array[0]['doc2weekend']<4) &&  ( weekday1==6 || weekday1 ==7) && ( weekday2!=6 && weekday2!=7) ){
-                    dhtmlx.message({ type:"error", text:array[0]['doc2']+"醫生假日班不得少於4" });
+                    var check = confirm(array[0]['doc2']+"醫生假日班不得少於4\n確定要換班嗎?");
+
+                    if(check==true){
+                       save_form();
+                    }
+                    else{
+                        alert("已取消");
+                    }
 
                 }
 
@@ -760,12 +772,27 @@
                 }
 
                  else if( (array[0]['doc1weekend']<=4) && ( weekday2==6 || weekday2 ==7)  && ( weekday1!=6 && weekday1!=7) ){
-                    dhtmlx.message({ type:"error", text:array[0]['doc1']+"醫生假日班不得少於4" });
-                    console.log("$week1"+weekday1);
-                    console.log("$week1"+weekday2);
+
+                    var check = confirm(array[0]['doc1']+"醫生假日班不得少於4\n確定要換班嗎?");
+
+                    if(check==true){
+                       save_form();
+                    }
+                    else{
+                        alert("已取消");
+                    }
+                    //dhtmlx.message({ type:"error", text:array[0]['docName']+"醫生假日班不得少於4" });
                 }
+                
                 else if((array[0]['doc2weekend']<4) &&  ( weekday1==6 || weekday1 ==7) && ( weekday2!=6 && weekday2!=7) ){
-                    dhtmlx.message({ type:"error", text:array[0]['doc2']+"醫生假日班不得少於4" });
+                    var check = confirm(array[0]['doc1']+"醫生假日班不得少於4\n確定要換班嗎?");
+
+                    if(check==true){
+                       save_form();
+                    }
+                    else{
+                        alert("已取消");
+                    }
 
                 }
                 else{
@@ -899,19 +926,40 @@
 
 
         function checkDoctorSchedule(id){
+            
             $.get('checkDoctorSchedule',{
                 scheduleID : id, // event id
                 date: document.getElementById("shiftDate").value //空格日期
     
             }, function(array){
-                if(array[0]!=0){
-                    //dhtmlx.message({ type:"error", text:"該天已有班" });
-                    refresh();
+                var scheduleID = id;
+                var classification = document.getElementById("shiftSessionID").value;
+                var date = document.getElementById("shiftDate").value;
+                var weekday = array[0]['weekDay'];
+                var weekDayInSchedule =array[0]['weekDayInSchedule'];
+
+                if(array[0]['count']!=0){
+                    dhtmlx.message({ type:"error", text:array[0]['docName']+"醫生"+array[0]['date']+"已有班" });
                 }
+
+                else if(array[0]['docWeekend']<=4 && (weekday != 6 && weekday != 7) && (weekDayInSchedule==6 || weekDayInSchedule==7)){
+                    var check= confirm(array[0]['docName']+"醫生假日班不得少於4\n確定要換班嗎?");
+                    if(check==true){
+                        updateSchedule(scheduleID,date,classification);
+                    }
+                    else{
+                        alert("已取消");
+                        refresh();
+                    }
+                     console.log(array[0]['docWeekend']);
+                     console.log(weekday);
+                }
+                else if(array[0]['date'] == array[0]['date']){
+                    updateSchedule(scheduleID,date,classification);
+                }
+                
+                
                 else {
-                    var date = document.getElementById("shiftDate").value;
-                    var scheduleID = id;
-                    var classification = document.getElementById("shiftSessionID").value;
                     updateSchedule(scheduleID,date,classification);
                     
                 }
@@ -934,8 +982,6 @@
             });
         }
 
-        
-
         function showScheduleInfo(date,section_id) {
              document.getElementById("shiftDate").value=date;
              document.getElementById("shiftSessionID").value=section_id;
@@ -957,13 +1003,38 @@
             }, function(array){
                 var weekday = array[0]['weekDay'];
                 if( array[0]['docWeekend'] <=4 && (weekday == 6 || weekday == 7)){
-                    dhtmlx.message({ type:"error", text:array[0]['docName']+"醫生假日班不得少於4" });
+                    var checkdelete = confirm(array[0]['docName']+"醫生假日班不得少於4\n確定要刪除嗎?");
+                    if(checkdelete==true){
+                        delete_confirm();
+                    }
+                    else{
+                        alert("已取消");
+                    }
+                }
+                if (array[0]['totalShift']<=array[0]['mustOnDuty']){
+                    var checkdelete = confirm(array[0]['docName']+"總班數不得小於"+array[0]['mustOnDuty']+"\n"+"確定要刪除嗎?");
+                    if(checkdelete==true){
+                        delete_confirm();
+                    }
+                    else{
+                        alert("已取消");
+                    }
+
                 }
                 else{
-                    delete_confirm();
+                    var date = document.getElementById("date1").options[document.getElementById('date1').selectedIndex].text;
+                    var r = confirm("是否刪除 " + array[0]['docName'] + " 在 " + date + " 的班?");
+                        if (r == true) {
+                            delete_confirm();
+                        } 
+                        else {
+                         alert("已取消");
+                     }
+                    
                 }
                
-        
+            console.log(array[0]['totalShift']);
+            console.log(array[0]['mustOnDuty']);
         });
 
         }
@@ -972,19 +1043,8 @@
                 scheduleID : document.getElementById('date1').value
             
             }, function(){
-                var doctor = document.getElementById("schID_1_doctor").options[document.getElementById('schID_1_doctor').selectedIndex].text;
-                var date = document.getElementById("date1").options[document.getElementById('date1').selectedIndex].text;
-                var r = confirm("是否刪除 " + doctor + " 在 " + date + " 的班?");
-                if (r == true) {
-                var event_id = scheduler.getState().lightbox_id;
-                scheduler.endLightbox(false, html("my_form"));
-                scheduler.deleteEvent(event_id);
                 alert("成功刪除");
                 refresh();
-            } else {
-                alert("已取消");
-            }
-        
         });
         }
     </script>
